refactor(world): WorldPage 컴포넌트 리팩토링 및 CC 감소
- WorldPage.tsx: CC 21/17 -> <10, 라인 수 819 -> 245 감소
- Components: CharacterNode, NetworkControls, NetworkDetailPanel 분리
- Hooks: useWorldGraph으로 로직 분리
- Utils: graphUtils.ts, constants.ts로 알고리즘 및 상수 분리
- 전체 리팩토링 목표 달성
