+++ refactoring.md
@@ -10,22 +10,27 @@
 npx eslint "src/**/*.{ts,tsx}" --rule '{"complexity": ["warn", 15]}'
 ```
 
-### 현재 CC >= 15 대상 (4개)
+### 현재 CC >= 15 대상 (3개)
 
 | #   | 파일                       | 함수                 | CC  | 상태                    |
 | --- | -------------------------- | -------------------- | --- | ----------------------- |
-| 1   | `TreeItem.tsx`             | TreeItem             | 33  | 🔧 훅 생성됨, 적용 필요 |
-| 2   | `CharacterDetailModal.tsx` | CharacterDetailModal | 20  | 🔧 훅 생성됨, 적용 필요 |
-| 3   | `WorldPage.tsx`            | CharacterNode        | 17  | 🚨 분석 필요            |
-| 4   | `WorldPage.tsx`            | Arrow function       | 21  | 🚨 분석 필요            |
+| 1   | `CharacterDetailModal.tsx` | CharacterDetailModal | 20  | 🔧 훅 생성됨, 적용 필요 |
+| 2   | `WorldPage.tsx`            | CharacterNode        | 17  | 🚨 분석 필요            |
+| 3   | `WorldPage.tsx`            | Arrow function       | 21  | 🚨 분석 필요            |
 
 ---
 
 ## ✅ 완료된 리팩토링
 
+### Phase 6: TreeItem 리팩토링 (CC 33 → 23)
+
+- **Hooks 분리**: `useTreeItem`, `useTreeItemMenu`
+- **UI 분리**: `TreeLines`
+- **결과**: 382줄 → 245줄, 로직과 UI 분리됨
+
 ### Phase 5: EditorPage 대규모 리팩토링 (CC 31 → 17)
 
-- **핸들러 분리**: `useEditorHandlers` (11개 핸들러)
+- **핸들러 분리**: `useEditorHandlers`
 - **이펙트 분리**: `useEditorEffects`, `useKeyboardSave`
 - **UI 분리**: `EditorToolbar`, `EditorContent`
 - **결과**: 939줄 → 475줄, 가독성 대폭 향상
@@ -42,13 +47,12 @@ npx eslint "src/**/*.{ts,tsx}" --rule '{"complexity": ["warn", 15]}'
 
 | 항목        | 시작 | 현재 |
 | ----------- | ---- | ---- |
-| CC>=15 경고 | 7개  | 4개  |
-| 신규 파일   | 0개  | 17개 |
+| CC>=15 경고 | 7개  | 3개  |
+| 신규 파일   | 0개  | 20개 |
 
 ---
 
 ## 🔧 남은 작업
 
-1. **TreeItem**: 생성된 훅을 메인 컴포넌트에 적용
-2. **CharacterDetailModal**: 생성된 훅을 완전 적용
-3. **WorldPage**: CharacterNode 및 Arrow function 분석 후 분리
+1. **CharacterDetailModal**: 생성된 훅을 완전 적용
+2. **WorldPage**: CharacterNode 및 Arrow function 분석 후 분리
